+++ core/src/test/java/com/jeanbarrossilva/orca/core/feed/profile/toot/style/styling/mention/SymbolMentionDelimiterTests.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling
+package com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.mention
 
 import kotlin.test.Test
 import kotlin.test.assertFalse
