+++ core/src/test/java/com/jeanbarrossilva/orca/core/feed/profile/toot/style/styling/mention/test/ColonMentionDelimiter.kt
@@ -1,6 +1,6 @@
-package com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.test
+package com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.mention.test
 
-import com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.Mention
+import com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.mention.Mention
 
 internal object ColonMentionDelimiter : Mention.Delimiter() {
     override fun getRegex(): Regex {
